chat changes are commited
