+++ package.json
@@ -13,20 +13,18 @@
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {    
+  "dependencies": {
     "body-parser": "^1.18.3",
     "bootstrap": "^4.1.2",
-    "eslint-config-standard": "^11.0.0",
     "express": "^4.16.3",
     "firebase": "^4.6.2",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "jquery-ui": "^1.12.1",
-    "material-components-web": "^0.38.0",
-    "material-design-lite": "^1.3.0",
+    "node-pre-gyp": "^0.10.3",
     "popper.js": "^1.14.3",
-    "webpack": "^4.16.3"
+    "socket.io": "^2.1.1"
   },
   "devDependencies": {
     "@material/drawer": "^0.38.0",
@@ -50,6 +48,8 @@
     "concurrently": "^3.6.1",
     "css-loader": "^1.0.0",
     "eslint": "^5.2.0",
+    "eslint-config-airbnb": "^17.0.0",
+    "eslint-config-airbnb-base": "^13.0.0",
     "eslint-config-standard": "^11.0.0",
     "eslint-config-webpack": "^1.2.5",
     "eslint-loader": "^2.1.0",
@@ -57,10 +57,7 @@
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^3.1.0",
-    "eslint-config-airbnb": "^17.0.0",
-    "eslint-config-airbnb-base": "^13.0.0",
     "file-loader": "^1.1.11",
-    "firebase": "^5.3.0",
     "glob": "^7.1.2",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
@@ -73,6 +70,7 @@
     "nodemon": "^1.18.3",
     "redux": "^4.0.0",
     "sass-loader": "^7.0.3",
+    "socket.io-client": "^2.1.1",
     "style-loader": "^0.21.0",
     "url-loader": "^1.0.1",
     "webpack": "^4.16.1",
